+++ Projeto 2-  com OO/Medicamento.cpp
@@ -4,7 +4,7 @@ Medicamento::Medicamento() {
     
 }
 
-Medicamento::Medicamento(std::string nome, int quantidade, int valorUnitario,
+Medicamento::Medicamento(std::string nome, int quantidade, double valorUnitario,
                             std::string dtVencimento, std::string nomeFabricante,
                             std::string dosagem, std::string administracao, std::string disponibilizacao)
                             : Insumo(nome, quantidade, valorUnitario, dtVencimento, nomeFabricante) {
